requirements.txt
imgaug==0.2.6
pandas==0.22.0
pafy==0.5.4
scipy==1.0.0
six==1.11.0
dill==0.2.7.1
matplotlib==2.2.3
hdbscan==0.8.13
lxml==4.1.1
scikit-learn==0.19.1
Pillow==6.2.1
opencv-python==3.4.0.12
numpy==1.14.0
Keras==2.1.4
tensorflow==1.4.0
h5py==2.7.1